pyproject.toml
[tool.poetry]
name = "dsa-python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "dsa_python"}]

[tool.poetry.dependencies]
python = "^3.10"
graphviz = "0.20.3"
recursion-visualiser = "1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
